+++ packages/widget_toolkit_otp/lib/widget_toolkit_otp.dart
@@ -1,17 +1,13 @@
 library widget_toolkit_otp;
 
-// Expose package specific enums which we would otherwise need to reimplement
-export 'package:pinput/pinput.dart'
-    show AndroidSmsAutofillMethod, PinAnimationType, HapticFeedbackType;
-
 export 'src/base/models/error/error_model.dart';
 export 'src/base/models/temporary_code_state.dart';
 export 'src/base/theme/sms_code_theme.dart';
 export 'src/base/utils/enums.dart';
 export 'src/base/utils/localized_strings.dart';
-export 'src/lib_countdown_widget/bloc/countdown_bloc.dart';
+export 'src/lib_countdown_widget/di/countdown_widget.dart';
 export 'src/lib_countdown_widget/services/countdown_service.dart';
-export 'src/lib_countdown_widget/widgets/countdown_widget.dart';
+export 'src/lib_countdown_widget/widgets/countdown_component.dart';
 export 'src/lib_resend_code_button/automated_resend_code_button.dart';
 export 'src/lib_sms_code_field/sms_code_field.dart';
 export 'src/lib_sms_code_field/sms_code_theme_configuration.dart';