+++ packages/widget_toolkit_otp/test/flutter_test_config.dart
@@ -1,6 +1,10 @@
 import 'dart:async';
+
 import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
+
+import 'helpers/widget_toolkit_file_comparator.dart';
 
 Future<void> testExecutable(FutureOr<void> Function() testMain) async =>
     GoldenToolkit.runWithConfiguration(
@@ -11,6 +15,12 @@ Future<void> testExecutable(FutureOr<void> Function() testMain) async =>
         return testMain();
       },
       config: GoldenToolkitConfiguration(
-        enableRealShadows: true,
-      ),
+          enableRealShadows: true,
+          fileNameFactory: (String name) {
+            final fileName = 'goldens/$name.png';
+
+            goldenFileComparator = WidgetToolkitFileComparator(fileName);
+
+            return fileName;
+          }),
     );