+++ packages/widget_toolkit_otp/CHANGELOG.md
@@ -1,2 +1,6 @@
-## [0.0.1]
-* Initial release
+## [0.0.1-dev1]
+* Initial release of the following libraries:
+  * SMS Code Field
+  * SMS Code Verification
+  * Resend Code Button
+  * Countdown Widget